build-publish-aws: maximize concurrency of GitHub jobs
The latest CI run for the build-publish-aws workflow took more than 5
hours, getting awkwardly close to the limit of 6 hours. Run all jobs of
the build strategy in parallel to maximize the amount of work that can
be done in parallel (mostly testing), and instead adjust the number of
concurrent upload tasks to fit into the service quota AWS has granted.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>