systemd-ephemeral: provide overlay mount for local CA certificates
The update-ca-certificates program expects to find local CA certificates
at /usr/local/share/ca-certificates. Since update-ca-certificates does
not make this path configurable and it is read-only on immutable images,
provide an overlay mount for it.

The certificates cannot be placed at /etc/ssl/certs (which is writable
on read-only images) instead. It is update-ca-certificates who places
links there. It does not use the directory as a source of certificates
to include in the certificate bundle it produces.

Related: https://github.com/chainguard-dev/internal-dev/issues/21930

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>