nvidia-cuda-crt-13.0: Port math_functions patch from nvcc (#4188)
The conflicting headers have moved from nvcc into this package, but they
are still impacted by the conflict with GLIBC 2.42+. Port the patch
over from nvidia-cuda-nvcc-12.9.

I have a test case for this, but it depends on a new wrapper package
that I have in development. I'll add a test along with that PR.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>