presubmit-reproducible - build for aarch64, call summary, and upload artifacts (#987)
We were running the same thing for aarch64 and x86_64 in this matrix.

The changes here
 1. validate a build of aarch64 'test/reproducible' image
 2. run image-summary on the output
 3. collect artifacts of those runs.

Note that the runners are always x86_64. apkoaas will qemu-system-x86_64
so we have to give it an x86_64 kernel and initrd even if though the
target content is aarch64.  I'm calling this out because it always
catches me.  It is how _all_ our aarch64 builds run.