Force etc-machine-id before network-pre.target for systemd-networkd (#37030)
I'm hoping this will fix
https://github.com/chainguard-dev/wolfi-vm/issues/971

There, Markus reported that
"Package not installed" (ENOPKG) is likely that machine-id
is not available at the time systemd dhcp executes.

There is nothing forcing systemd-networkd to occur after
machine-id.

This will cause that to happen, but arguably we should
just do this in systemd-networkd.service itself rather
than as a drop-in.