Better logging of qemu commands. (#992)
One fix of a confusing thing, and one better output. confusing: we were
printing output in qemu_cmmand.go as if it were being invoked, but it
was (yet).

prettier: quote output so that user can copy and paste it into shell.
Basically, cover it with quotes if it has non-shell-normal chars.