harden-vm: keep password-less enabled permanently in STIG profile
Commit c5b2bd23 ("harden-vm: keep password-less sudo enabled for STIG
profile (#36810)" disabled the sudo hardening in the STIG profile with
the intention of this change being temporary. The hardening was to be
enabled again once the test harness learned to set and use passwords for
privilege escalation.

In the meantime we noticed that "-slim" VM images, without agents like
cloud-init or Amazon SSM Agent, would have no way to set a password for
the images' default user and hence would not have a way to escalate
privileges. Therefore, keep disabling password-less sudo as a task for
the customer once all configuration requiring them to escalate
privileges has been done.

Issue: https://github.com/chainguard-dev/wolfi-vm/issues/969

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>