configs: add libpwquality for PAM module pam_pwquality
According to SBOM data, all images already have an implicit dependency
on linux-pam somewhere (e.g. via polkit or other packages). Make the
dependency explicit so the (lack of) variety is easier to reason about.

The DISA STIG requires controls to assess and maintain password
complexity. The control is implemented by means of the pam_pwquality PAM
module. Add the providing libpwquality package to all images, so that
PAM configurations that work on one (hardened) image will work as well
on another (not hardened) image.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>