feat(apm-server-9.0, fleet-server-9.2.yaml): Bump epochs to avoid customer repo sync conflicts (#37304)
https://github.com/chainguard-dev/extra-packages/pull/4207 PR removed and withdrew the packages from extras but the same verions of the packages
were alreayd in enterprise-packages. These were synced to customers repositories and because we had the same versioned packages in extras and
enterprise prior to withdrawal there was a conflict in this customer sync as the checksums did not match.

Even after withdrawal this conflict may continue to exist if the verison from extras was synced to the customer repos.

Bumping the epoch for the packages in enterprise ensures that the latest version of the package is synced to
customers repositories and the conflict is ignored.

Signed-off-by: philroche <phil.roche@chainguard.dev>