Update the ukify setup and re-enable aarch64 immutable ukify.  (#1023)
Things here:
 1. Add 'stampdalf_func' so we can call a function through stampdalf.
    this allows us to batch a bunch of things through stampdalf.
 2. turn ukify on for aarch64
 3. Be "arch aware" when calling ukify. Now we pass the sd-stub
    when calling ukify.  the stub must be installed in the root
    filesystem in order to handle this.  sd-stub comes from
    sytsemd-boot package.

This basically takes the loader/ configs and converts them
to ukify'd images with the same content. [1] covers the boot/ layout.
https://uapi-group.org/specifications/specs/boot_loader_specification

"Normally" (in fedora) the files in /EFI/Linux woudl be named like:

    <machine-id>-<uname-r>.efi

We don't have a machine-id at this point, so we're just using:

    <uname-r>.efi

This _does_ mean we can't have multiple entries per kernel-release there.
Thats not great, might have to be addressed.