tools: add stalemate tool to detect stale images and publication sadness
We have had situations in which published images were not updated in a
while and went stale, for example because:

  - Parts of our publishing workflows are asynchronous, like putting a
    new product version on the AWS Marketplace. A CI run may succeed,
    but nothing is monitoring for asynchronous failures after posting a
    new version for publishing.

  - A streak of CI failures, once in combination with our once-a-week
    publishing cadence for AWS, caused stale images even though the CI
    only intermittently went red.

Regardless of the reason, it is helpful to monitor any images published
remain fresh, and to alert on any images going stale, by looking at the
output state rather than only at the process (CI). To that end, add the
stalemate tool (to detect stale images and other stalemate situations,
like reaching the version limit on an AWS Marketplace product listing).

Currently, stalemate only covers the AWS Marketplace. I intend to check
the Azure Marketplace as well as our publishing channels (like S3 and
GCS buckets) in the future as well. A daily scheduled workflow will then
post a staleness report to #eng-vm-notifications for our on-call to
investigate.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>