amazon-eks-ami, amazon-eks-ami-fips: create nodeadm config drop-in dir
The amazon-eks-ami upstream release 20251023 includes a new feature [1]
that allows the chaining of nodeadm configuration providers, and the
nodeadm configuration to change during the boot process. The provided
systemd service unit for nodeadm-run.service sets /etc/eks/nodeadm.d as
a provider for config drop-ins. nodeadm then expects that directory to
actually exist and fails if it does not.

[1]: https://github.com/awslabs/amazon-eks-ami/pull/2427

Issue: https://github.com/chainguard-dev/wolfi-vm/issues/1045

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>