azure-eap-dev: add python3
We need a python3 command to be able to publish to the Marketplace. This is an
arbitrary and silly requirement, but it is a requirement that we have a `python`
or `python3` command that returns a version higher than 2.6. We used to get this
from the fact that cloud-init had a dependency on python-$ver instead of
python-$ver-base so it went unnoticed that this was missing from the image.