build-disks: bump memory required per VM in concurrency calculation
In a recent pre-submit runs I've seen builds be killed in what looks
like an OOM situation again. A rerun succeeded in both cases. We are
calculating the concurrency level based on the memory assigned to each
builder VM (4 GiB). This ignores potential overhead:

  - On my local Mac, I've seen the RSS to go actually up to slightly
    above 6 GiB per build.
  - On my Chainguard Workstation, I instrumented the build to run under
    the supervision of `/usr/bin/time -v` to report the RSS high-water
    mark and built all AWS images. This peaked at about 4.5 GiB
    committed.

Account for the overhead by bumping the memory requirement used in the
concurrency calculation to generous 5 GiB.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>