configs: convert hardened symlinks into regular configs
The hardened image variants have been symlinks to their regular
counterparts so far, with additional hardening steps only applied at
build time based on the variant's name. This has worked fine so far, but
always carried the risk of some tooling not properly dealing with the
symlinks or having to be aware of their existence.

Now, the symlinking approach caused problems: Recent changes to the
Makefile "yam --sort" the build configs. If executed in parallel on the
regular and hardened variants' configs, there is a window during which
one process can try to read from the file just truncated by the other
process and error out like this:

    unable to format "configs/aws-eks-1.32-dev-hardened/build.yaml": EOF

Convert the symlinks for the hardened variants into regular configs.
They 1:1 match the respective regular counterparts, but that does not
have to stay that way. Eventually, CUE will surely be able to help here.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>