configs: add dependency on ca-certificates to manage CA certs
Some images already have a transitive dependency on the ca-certificates
package that contains tooling to regenerate the system's certificate
bundle (which is provided by ca-certificates-bundle), thereby allowing
individual certificates to be trusted or distrusted. Managing the list
of trusted certificates is a basic requirement for all VM images, so do
add an explicit dependency on it to all build configurations.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>