docker-fips - restore missing dependencies. (#38477)
The previous commit inadvertantly dropped some dependencies from the 'docker' package.

The openssh-client dep would more correctly be put on docker-cli, but in the interest of not having changed things I'm adding it back here in 'docker' as it was before.

The ca-certificates package was also in 'docker', but really is needed for dockerd as it is likely to contact remote systems.

See also https://github.com/wolfi-dev/os/pull/71233