authentik: remediate CVE (#38504)
Remediate CVE-2025-64459
Import the upstream patch [1] that bumps django to 5.2.8.

It will also add a small patch from upstream.
The patch from upstream extends the number of characters allowed on
authentik URL's since django 5.2.8 enforces a lower number of characters
due to the fix for the CVE-64458 which is not applicable on Linux but
the character enforcement would affect authentik anyway.

https://github.com/goauthentik/authentik/commit/6a13b38740f92d8003e51fe2225513d1369e3e44

Signed-off-by: David Negreira <david.negreira@chainguard.dev>