fix(DCGM): Fix CUDART mismatch and dcgmproftester, stream with CUDA version (#38738)
DCGM was being built with CUDA 13.0, but shipped with CUDART 12.8 at
runtime. This bumps CUDART to 13.0 to resolve the mismatch

This also streams DCGM for CUDA 13.0, making situations like this in the
future much less likely. This also means DCGM will be bumped to newer
versions of CUDA as they're available

A few packages unfortunately depend on the "DCGM" provider and not the
"DCGM-cuda-X.Y" provider so I've retained the provider in our packaging.
I've set provider priority so that "DCGM" will always resolve to the
latest version of DCGM with the latest version of CUDA. We will need to
go back and update packages as needed

Oddly, a library required by dcgmproftester was split into the dev
package. Without it, dcgmproftester would be no-op. Looking at NVIDIA's
DCGM images, they do not remove it as suggested (it's included in
current tags for DCGM). So add it back to the main package

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>