vms-test: add new test group for SELinux
Add a test for SELinux to a new SELinux test group. The test checks
whether SELinux is enabled, the "targeted" policy is loaded, and no AVC
denials have been logged since the system started. At the moment, the
test accepts the system to be either in permissive and enforcing mode.

The SELinux test may over time move to the core group. For now, during
development, SELinux is an opt-in feature for individual images (like
qemu-selinux).

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>