libcrypto3-2.34: block hmac in fips mode
In this build of libcrypto3, update patches to the ones used in the
public openssl build of libcrypro3.

In this build of libcrypto3, verify that (unpatched) libcrypto works
with sha256-hmac in FIPS mode, and that this also works with the
patched libcrypto in FIPS mode too.

Separately will create a draft PR with the previous version of the
HMAC patch to show the failure.