vms-test: test-eks: wait in case of concurrent cluster creation (#1090)
After a cluster has been cleaned up or support for a new version of
Kubernetes has been added, CI runs needs to create a new cluster. Since
creating a new cluster takes up to half an hour, it is likely that
another job will run concurrently and find the cluster in the "CREATING"
state. Let this job wait for the cluster to become "ACTIVE" rather than
fail by not knowing how to deal with the state.

Last seen when introducing EKS 1.34 images to wolfi-vm:
https://github.com/chainguard-dev/wolfi-vm/actions/runs/19150922754/job/54740644544#step:31:273

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>