ec2-user: keep account locked but allow SSH login via PAM
Remove the systemd-sysusers drop-in that unlocks the ec2-user account.
Just keep the account locked as e.g. is the case on Amazon Linux 2023
as well. Still allow SSH login with a public key by setting "UsePAM yes"
in the sshd config, done via the new dependency on openssh-pam-config.

Commit c216b87a ("linky-user: integrate with pam and selinux") does
the same for the linky account (as currently used on QEMU and VMware
images) already.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>