py3-pygobject: fix overwriting upstream's pythoncapi-compat header with ours
Upstream dropped the patch we're applying in
https://github.com/GNOME/pygobject/commit/24206a1f0ee556c6126a8616253b28f6c098c835
as pythoncapi-compat upstream now ship a `meson.build` file.  This means
we need to properly check out the subproject before we copy our header
into place: the build system won't check out the upstream repo if the
directory already exists, and then the build fails due to the lack of
`meson.build`.