Compress qcow images with zstd rather than zlib (#1124)
The result of `make qcow-qemu-base-slim`
before and after:

   171781120 output/x86_64/qemu-base-slim/disk.qcow2.zlib
   165963264 output/x86_64/qemu-base-slim/disk.qcow2.zstd

So this is slightly better compression and ai tells me

> Zstd can be nearly twice as fast as zlib at decompressing data, even
> when both are set to achieve the same compression ratio.

So lets ship it.

There is a change in the qcow format, and some qemu's then would not be
able to use this directly. it seems that qemu 5.1 should have it, which
was released august 2020, but that doesn't mean that all qemu 5.1+ were
built with zstd.

That said, i say go with it.