verify openssl dynamicac linking for typst binary
this commit adds check so that we've sure that we're linking dynamically
against openssl when using cargo/build pipeline or doing a vanilla
cargo build against a project.

I've been here in the past as well because back in time I was packaging
lychee in wolfi and ended up here
https://github.com/lycheeverse/lychee/discussions/1376

`--features vendored-openssl` is a flag which comes from openssl crate
ref: https://docs.rs/openssl/0.10.75/openssl/#vendored

this is particularly useful to those who wants to opt in for static linking
but we don't want that in wolfi so we do some basic check in here to
verify that we're linking dynamically against openssl shared libraries.

In particular, we are doing two things below:
1. check for needed libraries at runtime and grep libssl or libcrypto in there
2. read the symbols table and grep for undefined openssl symbols in there.

fun fact and TIL: binutils ships all three of `objdump`, `nm` and `readelf`

down the road, I'll try to wrap up this logic in a pipeline so that it's
usable across other rust packages as well.