Makefile: only evaluate shell expansions once (#1162)
The PREFIX, COMMIT, and BUILD_TIMESTAMP variables can be passed in
externally. If not set (either as environment variable or Make
variable), they are defined as a shell expansion. However, with the "?="
operator they are not evaluated immediately, but defined as a macro with
deferred evaluation: The shell commands are invoked for each
use/expansion later on in the Makefile. This is harmless for PREFIX and
COMMIT, but can be problematic for BUILD_TIMESTAMP as the minute may
roll over between uses.

Ensure that all three variables are evaluated immediately.

Fixes: VMS-221

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>