vms-test: only package STIG Viewer artifact if it was to be created (#1180)
In the compliance/reports test, only package the STIG Viewer results
file if it was supposed to be created. At the moment, this is the only
the case for the GPOS STIG profile. For the other profiles, the attempt
to package the non-existing artifact would result in an error instead of
the file to be recorded in the artifacts.json which leads a later stage
of the pipeline to believe the test failed.

This caused failed jobs in last night's CI, e.g.
https://github.com/chainguard-dev/wolfi-vm/actions/runs/19916695026/job/57096833831

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>