notify-slack: prevent expansions on the message text (#1186)
Use single quotes for storing the message text in a variable to prevent
the shell from expanding variables and commands. This did not use to be
a problem, but:

- Commit 17e9adb1 ("notify-slack: accept message text input from a
file") set the trap by using double quotes
- Commit 416c0213 ("build-publish-*: embrace Markdown for the Slack
notifications (#1184)") walked into it when passing a message text
containing backticks

Expanding the message text can cause unintended modifications to the
text at best, or the action to fail at worst (e.g. command substitution
that fails). Then we're left without a Slack notification when we need
it the most.

Example: We should have received a notification for this failed build
job
(https://github.com/chainguard-dev/wolfi-vm/actions/runs/19998380348/job/57349678987),
but did not because the expansion failed and the notify-slack action
aborted.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>