selinux-policy: drop bogus container-selinux
As far as I can tell container-selinux policy is unused, also it is
unused in upstream builds as well.

And the fedora packaging is obsolete, and the container-selinux is
also unused there.

Instead we should package selinux-devel and then build
container-selinux separately.

Note that container-selinux policy is akward as it does `allow` on
`neverallow` things from the selinux-policy. I am not sure if or how
that works, if at all.

Also ./policy/users is a symlink to dist users already, thus no need
to copy it over the top.

Regenerating conf and compiling all contrib modules, is also broken as
many of them fail to compile.