runc-fips: remove all use of crypto from runc
According to the runc maintainers, runc does not support using
libcrypto.so from the host, and doing so in a context of FIPS builds is
unnecessary, since runc is not supposed to actually use any
cryptographic functions. [1]

Patch all vendored libraries to ensure they are not using the Golang
crypto package, so we can continue to compile runc-fips with the go-fips
toolchain and not defy expectations by dependencies or FIPS images using
runc-fips, while still fixing issues arising from using runc-fips to run
a container image without libcrypto.so. Also add a compile-time check
already used in wolfi-os to ensure that indeed the crypto package is not
being imported.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>