pipelines - daemon-check-output fix prepending shbang as documented. (#44005)
When reviewing https://github.com/chainguard-dev/enterprise-packages/pull/43860

I noticed that the "prepend '#!/bin/sh -ex'" that was documented
would not actually occur because the test to see if the document
already had a shbang would always return false.

So for 'setup' and 'post' scripts that did not have a shbang,
the documentation said they would get 'set -ex', but that was
not true.  They would not have '-e' or '-x'.