nginx-otel-mainline: Bump for rebuild (#44067)
Just bumping the epoch here to trigger a rebuild of the package.  I'm a
bit out of my depth here, but as far as I understand it this package is
statically linked to NGINX and performs a version check when the SO
module is loaded.  That means when we bump NGINX in the image we need to
rebuild this package with that version of NGINX.