postgresql-13/14/15: Fix Docker entrypoint
So up to now we were just curling files from Github and if we got a "404
Not Found", we'd just stick that into the file.  Now we will actually
fail if we get an HTTP failure code.

I also updated the Alpine versions too.  In the case of posgres 13 we
could only go to 3.22 and we also had to pin to the commit before the
files were removed.