linux: trust platform keyring for module signatures
Upstream rejected this: https://lkml.org/lkml/2025/6/2/451 for now.

They agree this is a genuine use-case in a VM where control of the keys in UEFI
is much easier, but say this compromises bare-metal security where UEFI vars contain
lots of keys from MSFT and hardware manufacturer that shouldn't be trusted by default.

For us, bare-metal is irrelevant and this makes sense.