fix(nemo): Include proper version of torchvision at runtime (#4772)
* fix(nemo): Include proper version of torchvision at runtime

Currently, the image depends on a virtual for torchvision-cuda12, which
points to torchvision for CUDA 12.3. NeMo currently uses CUDA 12.4 so
align the version of torchvision used at buildtime with the version used
at runtime by making it an explicit runtime dependency

Also unpin from 0.21.x. Newer versions of torchvision are compatible,
and have not introduced breaking changes

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(nemo): Bump resources

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

---------

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>