spring-boot: remove and withdraw package (#41666)
This package was introduced during a prospect engagement, but has never
ended up being used.  As spring-boot support multiple versions in
parallel, this package should really be version-streamed.  As it's
unused and therefore untested, we're instead withdrawing it, so we can
reintroduce it "properly" if we ever need to in future.  (That future
seems unlikely, with the advent of Libraries.)