consul-1.{20,21,22}: fix embeded consul binary versioning (#4770)
Wolfictl/syft detect a wrong version of consul 1.19.0 and this shows up
in the package sbom.
It seems that this is due to a version embedded in the UI of consul,
upstream does change this only when building their docker images, but
not in the source code itself so we should do it ourselves here in the
package.

More info and background:
https://github.com/chainguard-dev/internal-dev/issues/24929

Signed-off-by: David Negreira <david.negreira@chainguard.dev>