fix: Revert "pipelines - daemon-check-output fix prepending shbang as documented. (#44005)"
This reverts commit a29a91dec6ec3e8541cd864497dc852f0859c982.

The fix that was introduced is correct and is the desired behaviour
however, it's affected a number of packages and is blocking automation
from automagicking things away.

Unfortunately there's currently a spike in the queues so this is
adding workload.

We've agreed to temporarily revert and revisit in the new year

Signed-off-by: Ben Tasker <2900301+bentasker@users.noreply.github.com>