tritonserver-backend-pytorch-cuda-13.0: Stop vendoring binaries from other apks
The upstream build system expects to copy a bunch of files out of an
existing docker container image and vendor them into the backend's
subdirectory. This includes various libraries - some very large like
libtorch and libtorchvision .so's. Our existing patch was adjusting this
to work with Chainguard OS, copying files from other packages into this
one. But that means that if we fix a CVE (or any other issue) in those
packages, the copies we have here will not get updated.

Instead, let's disable the vendoring and add runtime dependencies
on the packages we need. Those files will no longer appear under
/opt/tritonserver/backend/, but the needed libraries should still
be resolvable. If we find that this causes issues, we might consider
adding symlinks to the files in those other packages.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>